# docs/openapi/control-api.yaml
openapi: 3.1.0
info:
  title: FlagForge Control API
  version: 1.0.0
  description: |
    Operator contract for tenant resources. Identity comes from an OIDC-issued bearer token; the
    organization is selected per request and is trusted only after an ACTIVE membership is found
    for that actor. Organization identifiers are never accepted from a request body.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/memberships:
    post:
      operationId: addMembership
      summary: Add a member to the caller's organization
      description: |
        The organization comes from the authenticated request, never from the body. Granting OWNER
        requires ORGANIZATION_MANAGE; every other role requires MEMBERSHIP_MANAGE.
      security:
        - operatorToken: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipRequest'
      responses:
        '201':
          description: Membership created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '400':
          $ref: '#/components/responses/InvalidTenantRequest'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
  /api/v1/memberships/role:
    put:
      operationId: changeMembershipRole
      summary: Change a member's role
      security:
        - operatorToken: []
      parameters:
        - $ref: '#/components/parameters/OrganizationHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MembershipRequest'
      responses:
        '200':
          description: Membership updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Membership'
        '401':
          $ref: '#/components/responses/AuthenticationRequired'
        '403':
          $ref: '#/components/responses/AccessDenied'
        '409':
          description: The change would leave the organization without an active OWNER.
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetail'
...
components:
  securitySchemes:
    operatorToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: OIDC-issued token whose subject identifies the operator.
  parameters:
    OrganizationHeader:
      name: X-FlagForge-Organization
      in: header
      required: true
      description: |
        Organization slug the request acts in. Accepted only when the authenticated actor holds an
        ACTIVE membership in it; otherwise the request is treated as unauthenticated, so an
        unknown organization and one the caller does not belong to cannot be told apart.
      schema:
        type: string
        pattern: '^[a-z0-9](?:[a-z0-9-]{0,61}[a-z0-9])?$'
...
  responses:
    AuthenticationRequired:
      description: |
        No usable identity. A missing token, a malformed or expired token, an unknown organization,
        and a missing membership all produce this one response.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
    AccessDenied:
      description: The actor is authenticated in this organization but its role lacks the permission.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
    InvalidTenantRequest:
      description: The request body is malformed, or a slug or key is already taken.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ProblemDetail'
...
  schemas:
...
    MembershipRequest:
      type: object
      required: [actorId]
      properties:
        actorId:
          type: string
          maxLength: 128
        role:
          type: string
          enum: [OWNER, ADMIN, DEVELOPER, VIEWER]
          description: Defaults to VIEWER when omitted on creation; required when changing a role.
    Membership:
      type: object
      required: [id, actorId, role, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        actorId:
          type: string
        role:
          type: string
          enum: [OWNER, ADMIN, DEVELOPER, VIEWER]
        status:
          type: string
          enum: [ACTIVE, SUSPENDED]
        createdAt:
          type: string
          format: date-time
...
    ProblemDetail:
      type: object
      description: RFC 9457 problem details.
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        errorCode:
          type: string
        correlationId:
          type: string